+++ mnist_snn/stdp.h
@@ -9,28 +9,28 @@
 */
 
 // Reinforcement learning curve
-float rl(int t)
+double rl(int t)
 {
     if (t > 0)
     {
-        return -A_plus * (expf(-(float)t / tau_plus) - stdp_offset);
+        return -A_plus * (exp(-(double)t / tau_plus) - stdp_offset);
     }
     else if (t <= 0)
     {
-        return A_minus * (expf((float)t / tau_minus) - stdp_offset);
+        return A_minus * (exp((double)t / tau_minus) - stdp_offset);
     }
 }
 
 // Update weights
-float update(float w, float del_w)
+double update(double w, double del_w)
 {
     if (del_w < 0.0)
     {
-        return w + sigma * del_w * powf((w - fabsf(w_min)), mu);
+        return w + sigma * del_w * pow((w - fabs(w_min)), mu);
     }
     else if (del_w > 0.0)
     {
-        return w + sigma * del_w * powf((w_max - w), mu);
+        return w + sigma * del_w * pow((w_max - w), mu);
     }
 }
